Quick code refactor (class and method names and package layout)
