+++ src/main/java/controllers/AgentsDataController.java
@@ -1,4 +1,4 @@
-package view;
+package controllers;
 
 
 import domain.AgentInfo;
