+++ src/main/java/main/Application.java
@@ -7,7 +7,7 @@ import org.springframework.data.mongodb.repository.config.EnableMongoRepositorie
 
 @SpringBootApplication
 @EnableMongoRepositories("dbmanagement")
-@ComponentScan({"dbmanagement", "view", "domain", "services"})
+@ComponentScan({"dbmanagement", "controllers", "domain", "services"})
 public class Application {
 
     public static void main(String[] args) {
